variable types change

+++ 0x13-bit_manipulation/1-print_binary.c
@@ -10,7 +10,7 @@
 
 void print_binary(unsigned long int n)
 {
-	int i = 0, copy = n, j = 0;
+	unsigned long int i = 0, copy = n, j = 0;
 
 	if (n == 0)
 	{
